+++ ovxlib/src/kernel/evis/upsample_evis.c
@@ -178,11 +178,11 @@ DEF_KERNEL_INITIALIZER(_upsample_initializer)
         input_fl = input_attr->dfp.fl;
         if (input_fl > 0)
         {
-            inputScale = 1.0f / (float) (1 << input_fl);
+            inputScale = 1.0f / (float) ((int64_t)1 << input_fl);
         }
         else
         {
-            inputScale = (float)(1 << -input_fl);
+            inputScale = (float)((int64_t)1 << -input_fl);
         }
     }
     else if( input_attr->quant == VSI_NN_KERNEL_QUANT_ASYMM )
@@ -196,11 +196,11 @@ DEF_KERNEL_INITIALIZER(_upsample_initializer)
         output_fl = output_attr->dfp.fl;
         if (output_fl > 0)
         {
-            outputScale = 1.0f / (float) (1 << output_fl);
+            outputScale = 1.0f / (float) ((int64_t)1 << output_fl);
         }
         else
         {
-            outputScale = (float)(1 << -output_fl);
+            outputScale = (float)((int64_t)1 << -output_fl);
         }
     }
     else if( output_attr->quant == VSI_NN_KERNEL_QUANT_ASYMM )
@@ -566,7 +566,7 @@ DEF_KERNEL_INITIALIZER(_upsample_initializer)
                 }
                 else
                 {
-                    uint32_t multiply       = (1 << (output_fl - input_fl));
+                    uint32_t multiply       = ((int64_t)1 << (output_fl - input_fl));
                     uint32_t i              = 0;
 
                     for (i = 8; i < 16; i++)
